backend-ai/requirements.txt
numpy
Pillow
flask
flask_cors
requests
python-dotenv
mysql-connector-python